deps: only install nvme-stas if there is a NVMe-oF controller
If a NVMe controller is present on the system, the storage configuration
will contain an object of type nvme_controller. Currently, the presence
of such object in the configuration leads to the installation of
nvme-cli and nvme-stas. However, the packages are not needed when the
controller uses the PCIe transport.

We now inspect the nvme_controller instances and install nvme-stas &
nvme-cli only if we detect a NVMe-oF controller.

LP: #2056730

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>