distro: fix handling of pre-installed OEM kernel
In the case of a pre-installed OEM kernel, we need to handle removals
differently.  Since Subiquity is managing OEM kernel installs, we need
that outside input to to know what the "before" state of the kernels is,
and the existing handling of ensure_one_kernel works well with that if
we just pass along that "before" state.

Have Subiquity feed the initial state (calculated before attempting to
install OEM kernels) will handle the OEM case - we start and finish with
one kernel so leave it alone - along with the two kernel case - we have
some other kernel pre-installed then Subiquity installs the OEM kernel,
and the remove list is that pre-installed kernel.