Do not read .flake8 configuration files when running py3-flake8
flake8, when invoked, will try to source configuration from a .flake8
file in the current directory. If it does not find it, it then looks in
the parent directory and so repeatedly until it finds a configuration
file (or gives up if no config file is found).

Since we do not have a .flake8 in the repository, the behavior of `tox
-e py3-flake8` is not consistent accross different environments.

Let's use the --isolated option so that the behavior is the same when
running tox -e py3-flake8 in CI and on local machines.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>