block_meta: ensure we don't use /dev/disk/by-path/XXX/by-uuid as volspec
Systems 256 introduced new symlinks in /dev/disk/by-path to avoid
potential UUID duplicates when the same image is written to multiple
storage units.

Such symlinks include:

 * /dev/disk/by-path/${path}/by-uuid/
 * /dev/disk/by-path/${path}/by-label/
 * /dev/disk/by-path/${path}/by-partnum/
 * /dev/disk/by-path/${path}/by-partuuid/
 * /dev/disk/by-path/${path}/by-partlabel/

When writing the fstab, we want curtin to specify the UUID of the
devices; rather than rely on a devlink which path can change when a
device gets removed or added.

To do so, we look for devlinks that match "by-uuid" and keep the first
match. Unfortunately, we now have two matches:

 * /dev/disk/by-uuid/${uuid}
 * /dev/disk/by-path/${path}/by-uuid/${uuid}

Ensure we use a stricter matching rule to avoid matching
/dev/disk/by-path/${path}/by-uuid/${uuid}.

LP: #2081256

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>