nvme-o-tcp: add function to adapt netplan config for dracut
During NVMe/TCP boot with NBFT, dracut renames the "main" network
interface to nbft0. This confuses netplan which expects the network
interface to be around with its original name. Furthermore, after
pivoting to the actual root FS, netplan (well, technically the backend
that netplan uses) wants to reapply the network configuration. This is a
problem with NVMe/TCP because the network interface should never be
brought down at runtime.

To address this issue, we now instruct netplan to match the
configuration to the HW address (i.e., the MAC address) instead of the
network interface name. We also mark the interface as "critical: true"
to ensure that the netplan backend does not bring it down.

LP: #2084012

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>