kdump: manual enable look for detection script
Make the manual enablement code path for kernel crash dumps check for the
enablement script. Previously we assumed any ProcessExecutionError
that resulted was from the script not existing, but it may be the case
that it fails for other reasons. Let's check if it exists first and just
not call it if it doesn't. This means we can let any
ProcessExecutionErrors bubble up instead of catching all of them.