Introduce a new boot builtin hook
Introduce a new hook for builtin_curthooks to configure the system to
boot. This currently uses grub but will also support extlinux in future.

The configuration key is 'boot'. It can be thought of as a rename of the
existing top-level 'grub' key, except that it allows one or more
bootloaders to be installed, using a new 'bootloaders' key. So the old
configuration:

    grub:
        install_devices:
            - /dev/vdb'

should now be expressed as:

    boot:
        bootloaders:
            - grub
        install_devices:
            - /dev/vdb'

The 'bootloaders' key is required. It does not have a default. This is
to ensure that each use of Curtin is converted with due care to which
bootloaders should be supported.

If 'boot' is missing, then a top-level 'grub' key is converted into
'boot', with all properties remaining the same except for the automatic
addition of:

    bootloaders:
        - grub

This maintains backwards compatibility with the old schema.

It would be quite nice to use an enum for the bootloaders.

Signed-off-by: Simon Glass <simon.glass@canonical.com>