Drop parser for install_grub
This is not used by the main program and it is probably not a good idea
to start using it. The functionality of this module is accessed by
calling the install_grub() function.

Remove this unused code.

Leave this in the commands/ subdir for now, but we could consider moving
it, e.g. into a new boot/ subdir.

Signed-off-by: Simon Glass <simon.glass@canonical.com>