doc: deb822 template fix
This is a follow-up from https://code.launchpad.net/~bryanfraschetti/curtin/+git/curtin/+merge/476864

As noted in the linked MP, the sources_list section is simply providing
a raw apt config, curtin doesn't process it. The `Key` field isn't documented
in sources.list(5), and didn't work for the customer.

The customer was able to use the Signed-By field with the ascii armored
key as shown here.

Signed-off-by: Wesley Hershberger <wesley.hershberger@canonical.com>