storage-config: don't crash when dealing with LDM or Windows dynamic disks
The Linux kernel has (partial) support for Windows dynamic disks (aka.
LDM) built on MSDOS ptables. If the kernel is built with
CONFIG_LDM_PARTITION (which is true in Ubuntu), it will detect partitions
that are not listed in the partition table itself, but in the LDM
database, stored in the last MiB of the disk.

In such a scenario, when Subiquity invokes curtin's storage-config using
data collected by probert, curtin fails with:

  RuntimeError: "couldn't find partition entry in table".

And this is expected since the partitions are listed in the LDM database
(which curtin can't read) and not in the partition table.

Instead of raising an error in this scenario, we now look if the disk
seems to have a LDM layout, and if it does, link the partition to the disk
but mark the ptable as "unsupported" instead of "msdos".

This allows subiquity to do its thing, knowing that a disk has an
"unsupported" layout.

(cherry picked from commit 59de02f54611932a827652c8b4e168c6260e0459)

LP: #2081778