storage-config: parse ptable field for RAID if specified
If the probe data reports a partition table on a RAID device, we need to
include it in the resulting storage action. Otherwise, the client needs to
guess the layout of partitions, and will most likely fail to do so.

With Subiquity, the absence of the ptable field leads to the following
issues:

 * failure to detect existing ESPs on GPT
 * upon creating a new partition on the RAID, the partition table is
   automatically changed to the default type (i.e., GPT except on s390x).

LP: #2098075

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>