storage-config: read the (GPT) partition name if present
If the PARTNAME property is present for a partition, its value will now be
passed to the partition fsobj in the partition_name field.

At least for GPT, the PARTNAME property only represents the partition name
correctly if it is made of ASCII characters. Otherwise, the value gets mangled
(the kernel drops the first byte of each UTF-16 character...). In certain
scenarios, it will even produce a value that can't be decoded as ASCII. If that
happens, the PARTNAME will be dropped by probert.

ID_PART_ENTRY_NAME is another property that we could use. It is set by libblkid
and does not suffer the same problems as PARTNAME. However, libblkid encodes
the partition name using blkid_encode_string. We'd need to come up with a
function that decodes the value using the same rules to obtain the original
partition name.

LP: #2099893

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>