storage-config: drop visible-on-ptable, unset ptable when relevant
When parsing a blockdev with DEVTYPE=partition, we look in the parent
blockdev for the partition table. Then, if we notice that the partition is
not actually associated with the partition table (which is only accepted
for Windows dynamic disks ATM), we used to add a "visible-in-ptable":
False attribute ; so that we don't try to retrieve properties from the
ptable later.

A simpler way it to unset the ptable variable. This has the same effect
without relying on an extra property.