block: fix string interpolation when building RuntimeError
The get_partition_sfdisk_info function means to raise a RuntimError if
the specified device is not found in the sfdisk output. However, the
string interpolation does not work because of missing parens. As a
consequence, a TypeError exception is raised instead of the
RuntimeError.

LP: #2059414

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>
(cherry picked from commit 9fc2aff85866f6a29a853fad652bd15eb0a3f9ee)