extlinux: Support a firmware-visible boot directory
When a separate boot partition is used, firmware such as U-Boot sees
this as a separate filesystem on a different device, since there is no
concept of a mount point in firmware.

This means that files must be loaded from the root directory of the boot
partition, rather than from '/boot'.

However a common case, which we need to support, is using a unified
filesystem. In this case the firmware must load files from the /boot
directory.

Provide support for setting this boot directory. This allows the correct
path to be used in extlinux.conf when loading the kernel and related
files.

This feature mirrors the _BOOT_DIRECTORY and _BOOT_PATH logic used by
the u-boot-menu package (in /usr/share/u-boot-menu/read-config).

For now, curthooks itself does not support providing this config.

Signed-off-by: Simon Glass <simon.glass@canonical.com>