extlinux: Split out processing into setup_extlinux()
The grub implementation uses a setup_grub() function which takes just
the boot configuration. Do the same for extlinux so we can add more
logic there, with access to the full config.

The test_install_extlinux() test actually calls install_extlinux() with
the wrong config. Fix that and tidy up some comments while we are here.

Signed-off-by: Simon Glass <simon.glass@canonical.com>