extlinux: Get the correct root device
The root device must be provided to the kernel using the 'append'
command in the extlinux.conf file. Add support for this.

Obtain the root device from block_meta, creating a new function for this
purpose.

Make use of select_configs() from the storage_config module, fixing up
an earlier, transitive import of this from block_meta

Signed-off-by: Simon Glass <simon.glass@canonical.com>