curthooks: Implement rescue mode in extlinux
Add a schema item controlling which alternatives to add to the extlinux
file. Update the implementation to respect this setting.

Use the 'rescue' terminology for this, since we may want to use
'recovery' for automatic recovery. It also matches the name used by
Ubuntu.

Separate out the BODY strings so we can mix and match them.

Add some comments to the tests as we have quite a few cases now. Also
comment the params fot build_content()

Signed-off-by: Simon Glass <simon.glass@canonical.com>