zfs: add dependency on cryptsetup for encrypted zpools
For a ZFS encrypted installation, the cryptsetup package must be
installed to the target system and the binary copied to the initramfs.
Failing to install the package results in an unbootable system.

Currently, we only install the cryptsetup package when the partitioning
config contains at least one "type: dm_crypt" element.

But not all installations have an explicit "type: dm_crypt" element.

When Subiquity runs a guided ZFS encrypted installation, the
partitioning config only contains a "type: dm_crypt" element if we have
room for an encrypted swap partition. While this is true in most
scenarios, installing to a small enough disk will make Subiquity skip
the swap partition.

We now also install crypsetup when the partitioning configuration
contains a "type: zpool" element with "encryption_style: luks_keystore".

LP: #2119293

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>