nvme-o-tcp: drop set-name directives from netplan config
If the nbft0 interface config contains a set-name directive, netplan will
attempt to honor it during boot. This leads to a temporary network loss, which
ends up fatal (i.e., it prevents booting) in a NVMe/TCP environment.

Make sure that we drop any set-name directive when configuring the critical
interface for NVMe/TCP booting.

LP: #2127072

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>